Updated and optimized data types

+++ Practice_team06/Models/Seat.cs
@@ -9,9 +9,9 @@ public partial class Seat
 
     public int HallId { get; set; }
 
-    public int RowNumber { get; set; }
+    public short RowNumber { get; set; }
 
-    public int SeatNumber { get; set; }
+    public short SeatNumber { get; set; }
 
     public decimal? PriceModifier { get; set; }
